Native: fixed vibe button navigation crushing android app
