+++ packages/app/features/profile/VibeButton.tsx
@@ -1,0 +1,17 @@
+import { LinkableButton, styled } from '@my/ui'
+
+const VibeButton = styled(LinkableButton, {
+  elevation: 1,
+  br: '$6',
+  ai: 'center',
+  jc: 'space-around',
+  w: 'auto',
+  maw: 100,
+  h: 'auto',
+  gap: '$2',
+  f: 1,
+  fd: 'column',
+  p: '$3.5',
+})
+
+export default VibeButton
