+++ packages/app/features/profile/screen.tsx
@@ -9,11 +9,9 @@ import {
   Card,
   H3,
   Image,
-  LinkableButton,
   Paragraph,
   Spinner,
   Stack,
-  styled,
   Text,
   useMedia,
   useThemeName,
@@ -36,6 +34,7 @@ import { useHoverStyles } from 'app/utils/useHoverStyles'
 import { Linking, Platform, Pressable } from 'react-native'
 import ProfileSendButton from 'app/features/profile/ProfileSendButton'
 import ViewHistoryButton from 'app/features/profile/ViewHistoryButton'
+import VibeButton from 'app/features/profile/VibeButton'
 
 interface ProfileScreenProps {
   sendid?: number | null
@@ -230,20 +229,6 @@ export function ProfileScreen({ sendid: propSendid }: ProfileScreenProps) {
     </YStack>
   )
 }
-
-const VibeButton = styled(LinkableButton, {
-  elevation: 1,
-  br: '$6',
-  ai: 'center',
-  jc: 'space-around',
-  w: 'auto',
-  maw: 100,
-  h: 'auto',
-  gap: '$2',
-  f: 1,
-  fd: 'column',
-  p: '$3.5',
-})
 
 const Vibe = ({
   amount,
